Add files via upload
IThis youtube clone webpage simply displays the use essesntial css features, i.e The grid/flex-box display layouts, position relative/absolute, the z-index value, and use of the @media. 
I have used the grid property for the main video priview page and the flex-box for the header and the sidebar of the youtubeclone page. check it  through